+++ nglview/utils.py
@@ -20,76 +20,6 @@ def get_name(obj, kwargs):
         name = name.split()[0].strip('<')
     return name
 
-def _set_ipython_cell():
-    cm = """
-    var setIPythonLikeCell = function(){
-        //var cell = Jupyter.notebook.insert_cell_at_bottom();
-        var cell = Jupyter.notebook.get_selected_cell();
-        var $el = cell.element;
-        cell.set_text('');
-        $el.css({'background': '#e5f5f9'});
-
-        var handler = function(event) {
-            var selected_cell = Jupyter.notebook.get_selected_cell();
-            if (selected_cell.cell_id === cell.cell_id){
-                selected_cell.execute();
-                selected_cell.set_text('');
-            }
-            return false;
-        };
-
-        var action = {
-            help: 'run cell',
-            help_index: 'zz',
-            handler: handler
-        };
-
-        Jupyter.keyboard_manager.edit_shortcuts.add_shortcut('enter', action); 
-    };
-
-    setIPythonLikeCell()
-    """
-    from IPython.display import display, Javascript, HTML
-    display(Javascript(cm))
-
-def ngl_demo(width=400, height=400):
-    """make a viewport, create stage object and populate NGL namespace
-    """
-    from IPython.display import display, Javascript, HTML
-
-    command = """
-    <div id='viewport'></div>
-    <script>
-        $('#viewport').width(%s).height(%s);
-    </script>
-    """ % (width, height)
-
-    command2 = """
-    <script>
-        var NGL = require('nbextensions/nglview/widget_ngl').NGL;
-        var stage = new NGL.Stage('viewport')
-    </script>
-    """
-
-    display(HTML(command))
-    display(HTML(command2))
-
-def init_js_funcs():
-    """print
-    """
-    from IPython.display import display, Javascript, HTML
-
-    command = """
-    <script>
-        var print = function(x){
-            for (var i in x){
-                console.log(i)
-            }
-        }
-    </script>
-    """
-    display(HTML(command))
-
 def get_repr_names_from_dict(repr_dict, component):
     """
